chore (mongo): Common ConfigurationSchemaDocument and FieldDefinitionDocument
