+++ persistence-spring-mongo/src/main/java/io/naryo/infrastructure/configuration/persistence/document/broadcaster/configuration/BroadcasterConfigurationDocument.java
@@ -1,14 +1,15 @@
 package io.naryo.infrastructure.configuration.persistence.document.broadcaster.configuration;
 
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 
 import com.mongodb.lang.Nullable;
 import io.naryo.application.configuration.source.definition.ConfigurationSchema;
-import io.naryo.application.configuration.source.definition.FieldDefinition;
 import io.naryo.application.configuration.source.model.broadcaster.configuration.BroadcasterCacheConfigurationDescriptor;
 import io.naryo.application.configuration.source.model.broadcaster.configuration.BroadcasterConfigurationDescriptor;
 import io.naryo.domain.broadcaster.BroadcasterType;
+import io.naryo.infrastructure.configuration.persistence.document.common.ConfigurationSchemaDocument;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -52,7 +53,7 @@ public final class BroadcasterConfigurationDocument implements BroadcasterConfig
 
     @Override
     public Optional<ConfigurationSchema> getPropertiesSchema() {
-        return Optional.ofNullable(fromDocument(this.propertiesSchema));
+        return Optional.ofNullable(ConfigurationSchemaDocument.fromDocument(this.propertiesSchema));
     }
 
     @Override
@@ -67,47 +68,6 @@ public final class BroadcasterConfigurationDocument implements BroadcasterConfig
 
     @Override
     public void setPropertiesSchema(ConfigurationSchema propertiesSchema) {
-        this.propertiesSchema = toDocument(propertiesSchema);
-    }
-
-    private static ConfigurationSchema fromDocument(ConfigurationSchemaDocument doc) {
-        if (doc == null) return null;
-
-        List<FieldDefinition> fields =
-                doc.getFields().stream()
-                        .map(
-                                fd -> {
-                                    try {
-                                        Class<?> clazz = Class.forName(fd.getTypeName());
-                                        return new FieldDefinition(
-                                                fd.getName(),
-                                                clazz,
-                                                fd.isRequired(),
-                                                fd.getDefaultValue());
-                                    } catch (ClassNotFoundException e) {
-                                        throw new RuntimeException(
-                                                "Class not found: " + fd.getTypeName(), e);
-                                    }
-                                })
-                        .collect(Collectors.toList());
-
-        return new ConfigurationSchema(doc.getType(), fields);
-    }
-
-    private static ConfigurationSchemaDocument toDocument(ConfigurationSchema schema) {
-        if (schema == null) return null;
-
-        List<FieldDefinitionDocument> fieldDocs =
-                schema.fields().stream()
-                        .map(
-                                f ->
-                                        new FieldDefinitionDocument(
-                                                f.name(),
-                                                f.type().getName(),
-                                                f.required(),
-                                                f.defaultValue()))
-                        .collect(Collectors.toList());
-
-        return new ConfigurationSchemaDocument(schema.type(), fieldDocs);
+        this.propertiesSchema = ConfigurationSchemaDocument.toDocument(propertiesSchema);
     }
 }
