+++ persistence-spring-mongo/src/main/java/io/naryo/infrastructure/configuration/persistence/document/common/FieldDefinitionDocument.java
@@ -1,4 +1,4 @@
-package io.naryo.infrastructure.configuration.persistence.document.broadcaster.configuration;
+package io.naryo.infrastructure.configuration.persistence.document.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
