+++ persistence-spring-mongo/src/main/java/io/naryo/infrastructure/configuration/persistence/document/event/store/EventStoreConfigurationPropertiesDocument.java
@@ -4,6 +4,7 @@ import io.naryo.application.configuration.source.definition.ConfigurationSchema;
 import io.naryo.application.configuration.source.model.event.EventStoreConfigurationDescriptor;
 import io.naryo.domain.configuration.eventstore.EventStoreStrategy;
 import io.naryo.domain.configuration.eventstore.EventStoreType;
+import io.naryo.infrastructure.configuration.persistence.document.common.ConfigurationSchemaDocument;
 import jakarta.annotation.Nullable;
 import jakarta.validation.constraints.NotNull;
 import lombok.Setter;
@@ -14,17 +15,17 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
-@Document(collection = "event_store")
+@Document(collection = "event_stores")
 @Setter
 public abstract class EventStoreConfigurationPropertiesDocument implements EventStoreConfigurationDescriptor {
     private @MongoId String nodeId;
     private @NotNull EventStoreType type;
     private @NotNull EventStoreStrategy strategy;
     private Map<String, Object> additionalProperties;
-    private @Nullable ConfigurationSchema propertiesSchema;
+    private @Nullable ConfigurationSchemaDocument propertiesSchema;
 
     public EventStoreConfigurationPropertiesDocument(String nodeId, EventStoreType type, EventStoreStrategy strategy, @Nullable Map<String, Object> additionalProperties,
-                                                     @Nullable ConfigurationSchema propertiesSchema) {
+                                                     @Nullable ConfigurationSchemaDocument propertiesSchema) {
         this.nodeId = nodeId;
         this.type = type;
         this.strategy = strategy;
@@ -34,7 +35,7 @@ public abstract class EventStoreConfigurationPropertiesDocument implements Event
 
     @Override
     public UUID getNodeId() {
-        return UUID.fromString(this.nodeId);
+        return UUID.fromString(nodeId);
     }
 
     @Override
@@ -48,12 +49,22 @@ public abstract class EventStoreConfigurationPropertiesDocument implements Event
     }
 
     @Override
-    public Optional<Map<String, Object>> getAdditionalProperties() {
-        return Optional.ofNullable(additionalProperties);
+    public Map<String, Object> getAdditionalProperties() {
+        return additionalProperties;
+    }
+
+    @Override
+    public void setAdditionalProperties(Map<String, Object> additionalProperties) {
+        this.additionalProperties = additionalProperties;
     }
 
     @Override
     public Optional<ConfigurationSchema> getPropertiesSchema() {
-        return Optional.ofNullable(propertiesSchema);
+        return Optional.ofNullable(ConfigurationSchemaDocument.fromDocument(propertiesSchema));
+    }
+
+    @Override
+    public void setPropertiesSchema(ConfigurationSchema propertiesSchema) {
+        this.propertiesSchema = ConfigurationSchemaDocument.toDocument(propertiesSchema);
     }
 }
