+++ persistence-spring-mongo/src/main/java/io/naryo/infrastructure/configuration/persistence/document/event/store/block/database/DatabaseBlockEventStoreConfigurationPropertiesDocument.java
@@ -3,35 +3,37 @@ package io.naryo.infrastructure.configuration.persistence.document.event.store.b
 import java.util.Map;
 import java.util.Set;
 
-import io.naryo.application.configuration.source.definition.ConfigurationSchema;
 import io.naryo.application.configuration.source.model.event.DatabaseBlockEventStoreConfigurationDescriptor;
 import io.naryo.domain.configuration.eventstore.EventStoreType;
 import io.naryo.domain.configuration.eventstore.database.DatabaseEngine;
+import io.naryo.infrastructure.configuration.persistence.document.common.ConfigurationSchemaDocument;
 import io.naryo.infrastructure.configuration.persistence.document.event.store.block.BlockEventStoreConfigurationPropertiesDocument;
 import io.naryo.infrastructure.configuration.persistence.document.event.store.block.EventStoreTargetPropertiesDocument;
 import jakarta.validation.constraints.NotNull;
 import lombok.Setter;
 import org.jetbrains.annotations.Nullable;
+import org.springframework.data.annotation.TypeAlias;
 
 @Setter
+@TypeAlias("database_event_store")
 public final class DatabaseBlockEventStoreConfigurationPropertiesDocument
         extends BlockEventStoreConfigurationPropertiesDocument
         implements DatabaseBlockEventStoreConfigurationDescriptor {
 
-    private DatabaseEngine engine;
+    private String engine;
 
     public DatabaseBlockEventStoreConfigurationPropertiesDocument(
             @NotNull String nodeId,
             Set<EventStoreTargetPropertiesDocument> targets,
             Map<String, Object> additionalProperties,
-            @Nullable ConfigurationSchema propertiesSchema,
-            @NotNull DatabaseEngine engine) {
+            @Nullable ConfigurationSchemaDocument propertiesSchema,
+            @NotNull String engine) {
         super(nodeId, EventStoreType.DATABASE, targets, additionalProperties, propertiesSchema);
         this.engine = engine;
     }
 
     @Override
     public DatabaseEngine getEngine() {
-        return engine;
+        return () -> engine.toLowerCase();
     }
 }
